Refactor: split process.rs
