+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "kfs"
-version = "0.0.2"
+version = "0.0.3"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
