+++ src/interrupt/hw.rs
@@ -1,1 +1,2 @@
+pub mod apic_timer;
 pub mod keyboard;
